benchmarks/office_cert_temporal_poc/pyproject.toml
[project]
name = "office-cert-temporal-poc"
version = "0.1.0"
description = "PoC tooling for temporal office-certification benchmark tasks."
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "msoffcrypto-tool>=5.4.2",
  "olefile>=0.47",
  "openpyxl>=3.1.5",
  "pandas>=2.2.3",
  "pdf2image>=1.17.0",
  "python-docx>=1.1.2",
  "pyyaml>=6.0.2",
  "requests>=2.32.0",
]

[tool.uv]
package = false
